proxy-fetcher/requirements.txt
cassandra-driver==3.29.2
requests==2.32.3
